:truck: moved content checking actions into the email module.
